requirements.txt
python-dateutil==2.8.2
beautifulsoup4==4.12.2
requests==2.31.0
flask==3.0.0
gunicorn==23.0.0
selenium==4.16.0
webdriver-manager==4.0.1

# Authentication and database dependencies
flask-login==0.6.3
python-dotenv==1.0.0
cryptography==41.0.7
email-validator==2.1.0
